pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "trace-regression-harness"
version = "0.1.0"
description = "Policy assertions for agent traces — the quality gate for AI runs"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-asyncio>=0.23.0"]

[project.scripts]
trace-check = "cli.main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
